+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * tag provisioned s3 buckets with the opsworks stack name
 * Fix a couple of issues that resulted in modifications to files as a
   side-effect of running `bin/setup`.
+* prompt user to consider using a local vagrant cluster
 
 ## 1.7.1 - 7/20/2016